requirements.txt
faster-whisper
flask
flask-cors
requests
anthropic
python-dotenv
